Add nostrum explicitly to dialyzer’s apps.
In order to ease setup for other devs who just want to get the phoenix
app working and make changes without fiddling with a Discord bot 
account, we don’t start the bot unless it is specifically configured. 
Dialyzer was ignoring the dep but complaining that there is code in the
app that calls it in the static analysis.