Add a full static HTML home page for the GEM event
I think the styling here would be a good homepage for us, but in an
effort to get this updated quickly without rewriting too much it's just
static for now.

In the future we could integrate with the database or look at going to
a fully static site generator like Hugo.